+++ trunk/SHM/src/main/java/ua/com/manometer/dao/invoice/PaymentDAO.java
@@ -9,6 +9,6 @@ public interface PaymentDAO {
 
 	public List<Payment> listPayment();
 
-	public void removePayment(Long id);
+	public void removePayment(Integer id);
 
 }
