+++ trunk/SHM/src/main/java/ua/com/manometer/service/invoice/InvoiceFilterService.java
@@ -10,6 +10,6 @@ public interface InvoiceFilterService {
 
 	public List<InvoiceFilter> listInvoiceFilter();
 
-	public void removeInvoiceFilter(Long id);
+	public void removeInvoiceFilter(Integer id);
 
 }
